Add files via upload
This code is the back-end (PHP) code for a spell check program. It uses the Levenshtein distance algorithm to compute the distance between the incorrectly spelled word and all other words in a dictionary; a suggested (correct) spellings list appears and the user has the option of selecting any of the recommendations to replace the incorrectly spelled word in question.